insta-crawl/requirements.txt
python-instagram==1.3.2